requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
sqlalchemy
pymysql
websockets
redis